Now supports diffrerent types in powershell

+++ init_template.ps1
@@ -2,7 +2,6 @@ $type=$args[0]
 $app_name=$args[1]
 $open=$args[2]
 
-
 if($app_name -ne '.' -And $app_name -ne '' ){
     mkdir $app_name;
     Set-Location $app_name;
@@ -14,33 +13,37 @@ if($open -eq 'open'){
 
 git init;
 
-switch ($type) {
-    'react' { 
-        npm create vite@latest -- --template react;
+if($type -ne 'node' -And $type -ne 'new' -And $type -ne ''){
+    if($type.split('-').length -ne 2){
+        npm create vite@latest -- --template $type.split('-')[0] .;
         npm i;
         Copy-Item $PSScriptRoot\.gitignore  .
     }
-    'react-node' {
+    else{
         mkdir frontend;
         Set-Location frontend;
-        npm create vite@latest -- --template react .;
+        npm create vite@latest -- --template $type .;
         npm i;
         Remove-Item .gitignore;
         Set-Location ..;
         mkdir backend;
         Set-Location backend;
         Copy-Item $PSScriptRoot\node\* -Destination . -Recurse;
-        npm i express dotenv;
+        npm i express dotenv express-async-handler;
         Set-Location ..;
         Copy-Item $PSScriptRoot\.gitignore .;
     }
-    'node'{
-        Copy-Item $PSScriptRoot\node\* . -Recurse;
-        Copy-Item $PSScriptRoot\.gitignore  .;
-        npm i express dotenv;
+}
+else{
+    switch ($type) {
+        'node'{
+            Copy-Item $PSScriptRoot\node\* . -Recurse;
+            Copy-Item $PSScriptRoot\.gitignore  .;
+            npm i express dotenv express-async-handler;
+        }
+        'new'{
+            Copy-Item $PSScriptRoot\vanilla\* .;
+         }
+        Default: { Write-Output "Invalid-Option" }
     }
-    'new'{
-        Copy-Item $PSScriptRoot\vanilla\* .;
-     }
-    Default { Write-Output 'Invalid Option' }
 }
